% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{radar_wqi}
\alias{radar_wqi}
\title{Radar chart for visualizating relative contributions of water quality parameters to CCME index}
\usage{
radar_wqi(data)
}
\arguments{
\item{data}{A data object (wqitab) from calc_wqi().}
}
\description{
Radar chart for visualizating relative contributions of water quality parameters to CCME index
}
\examples{
radar_wqi(calc_wqi(ccme)$wqitab)
}
