\name{wqifun}
\alias{wqifun}

\title{
WQI
}
\description{the function calculates the water quality indices using the unweighted arthematic WQI formula
}
\usage{
wqifun(x, ...)
}

\arguments{
  \item{x}{ matrix of normalized water quality parameter values}
  \item{\dots}{other arguments as may be required for the function.}
 }


\value{
  \item{wqi}{ vector where the calculated Water quality index values are stored}
  }


\author{Jigeeshu Joshi <j_josh01@uni.muenster.de>}





\keyword{ wqi}

