% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{wqindex}
\alias{wqindex-package}
\alias{wqindex}
\title{Water Quality Index Calculation}
\description{
The wqindex package calculates water quality thresholds and indices for
freshwater life in British Columbia. To the best of the
package creator's knowledge the limits correspond to the
approved Provincial thresholds on the date of
release but wqbc comes with ABSOLUTELY NO WARRANTY. For further
information on the use of the wqbc package please type
\code{vignette("wqindex")} at the R console.
}
\examples{
\dontrun{
 demo(ccme)
 demo(dummy)
 demo(fraser)
}
}
\seealso{
\code{\link{calc_wqi}}
and \code{\link{plot_wqis}}
}
\author{
\strong{Maintainer}: Karly Harker \email{karly.harker@gov.bc.ca} [contributor]

Authors:
\itemize{
  \item Joe Thorley \email{joe@poissonconsulting.ca} (Primary author) [contractor]
  \item Colin Millar \email{colinpmillar@gmail.com} [contractor]
  \item Andy Teucher \email{andy.teucher@gov.bc.ca}
}

Other contributors:
\itemize{
  \item Wendy Wang [contributor, contractor]
  \item Robyn Irvine [contributor, contractor]
  \item Province of British Columbia [copyright holder]
}

}
